Initial version of attachfiles plugin
attachfiles copies/moves/links/hardlinks files that should go together
with your music. It's configured with:

attachfiles:
    files = *.jpg info covers

so when you import an album that has jpg files next to the music, the
files will be copied/moved/... along with the music. If the pattern
is a directory, the directory is copied/moved/... recursively