Implement attachment removals, simplifications and two new options
Added implementation of item_removed that actually removes the
attachments if the directory is being emptied. The path
string handling has been simplified, as has been the item_removed
code that was already implemented.

Two new options have been added and implemented:
'allow_multiple_copies' (default False) : When True, allows the plugin
to copy the same attachment files to multiple directories if the music
files are being spread to different places. By default, the attachment
files/directories are only copied to the location of the first music
file parsed in each directory.
'allow_remove': (default True) : When True, if the user removes music
from disk, the files/directories that match the attachment patterns
are removed from disk too if that means the directory will get empty
after the operation. If False, the attachments are never removed by
the plugin.