Use unique name when attaching a single file or symlinking a directory
In the *_file_or_directory functions, if `src` is a file and `tgt` already
exists, use a unique name as the recursive functions do. Also,
link_file_or_directory never recurses. It just tries to create a symlink
if `src` is a file or a directory, and if `tgt` already exists, uses
a unique name.